re-organized structure for autocorrect
