+++ src/microbit/__model/constants.py
@@ -64,7 +64,7 @@ IMAGE_PATTERNS = {
     "SWORD": "00900:00900:00900:09990:00900:",
     "GIRAFFE": "99000:09000:09000:09990:09090:",
     "SKULL": "09990:90909:99999:09990:09990:",
-    "UMBRELLA": "09990:99999:00900:90900:09900:",
+    "U__mbRELLA": "09990:99999:00900:90900:09900:",
     "SNAKE": "99000:99099:09090:09990:00000:",
 }
